+++ core-ng/src/main/java/core/framework/impl/web/WebContextImpl.java
@@ -1,5 +1,7 @@
 package core.framework.impl.web;
 
+import core.framework.api.util.Maps;
+import core.framework.api.web.Request;
 import core.framework.api.web.WebContext;
 
 import java.util.Map;
@@ -8,10 +10,31 @@ import java.util.Map;
  * @author neo
  */
 public class WebContextImpl implements WebContext {
-    final ThreadLocal<Map<String, Object>> context = new ThreadLocal<>();
+    private final ThreadLocal<Map<String, Object>> context = new ThreadLocal<>();
+    private final ThreadLocal<Request> request = new ThreadLocal<>();
 
+    @Override
     @SuppressWarnings("unchecked")
     public <T> T get(String key) {
         return (T) context.get().get(key);
+    }
+
+    @Override
+    public Request request() {
+        return request.get();
+    }
+
+    void put(String key, Object value) {
+        context.get().put(key, value);
+    }
+
+    void initialize(Request request) {
+        context.set(Maps.newHashMap());
+        this.request.set(request);
+    }
+
+    void cleanup() {
+        context.remove();
+        request.remove();
     }
 }